requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==4.2
django-filter==24.2
django-pytest==0.2.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
flake8==7.0.0
freezegun==1.5.1
iniconfig==2.0.0
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pyflakes==3.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.1
tzdata==2024.1
django-jazzmin==2.4.8